moves tests to utils
