+++ packages/web/app/lib/web3/light-wallet/lightWalletUtils.spec.ts
