+++ packages/web/app/test-e2e/obsolete/generateRandomSeedAndAddress.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from "chai";
 
 import { isAddressValid } from "../../modules/web3/utils";
+import { DEFAULT_HD_PATH } from "./../utils/constants";
 import { generateRandomSeedAndAddress } from "./generateRandomSeedAndAddress";
-import { DEFAULT_HD_PATH } from "../utils/userHelpers";
 
 describe("generateRandomSeedAndAddress", () => {
   it("generates random seed and password", async () => {
