+++ packages/web/app/test-e2e/obsolete/generateSeedAndAddress.spec.ts
