+++ packages/web/package.json
@@ -49,8 +49,7 @@
     "test:e2e:cypress:io": "cypress run --env Tags=$(cat ./e2e-test-tags) --config testFiles=\"**/*.spec.e2e.io.ts\" --record --parallel --group remoteIOTests --ci-build-id $CIRCLE_BUILD_NUM$(echo node)",
     "test:e2e:cypress:vis-reg": "echo build id $CIRCLE_BUILD_NUM$(echo visReg) && cypress run --browser chrome --config testFiles=\"**/*.spec.e2e.visreg.ts\" --record --parallel --group io --ci-build-id $CIRCLE_BUILD_NUM$(echo visReg)",
     "test:e2e:dev": "cypress open --config testFiles=\"**/*.spec.e2e.*\"",
-    "test:node": "TS_NODE_PROJECT=tsconfig.spec.json mocha --require ts-node/register/transpile-only --require ./test/setupTests.ts './app/**/*.spec.node.ts?(x)'",
-    "test": "yarn test:browser && yarn test:node",
+    "test": "yarn test:browser",
     "tsc": "concurrently -p \"[{name}]\" -n \"TSC:APP,TSC:E2E\" -c \"bgYellow.bold,bgMagenta.bold\" \"yarn tsc:app\" \"yarn tsc:e2e\"",
     "tsc:app": "tsc --noEmit --project ./tsconfig.json",
     "tsc:e2e": "tsc --noEmit --project ./tsconfig.e2e.json",
